+++ starlight-core/src/main/java/org/atcraftmc/starlight/framework/pack/MultiPackageProvider.java
@@ -3,6 +3,7 @@ package org.atcraftmc.starlight.framework.pack;
 import me.gb2022.gluon.pack.ApplicationPackage;
 import org.atcraftmc.starlight.Starlight;
 import org.atcraftmc.starlight.framework.SLPluginConcept;
+import org.atcraftmc.starlight.util.EarlyLoadingManager;
 import org.atcraftmc.starlight.util.ProductMetadata;
 import org.atcraftmc.starlight.util.dependency.LibraryManager;
 import org.bukkit.plugin.java.JavaPlugin;
@@ -28,6 +29,11 @@ public abstract class MultiPackageProvider extends JavaPlugin implements Package
             return;
         }
         this.packages = createPackages();
+    }
+
+    @Override
+    public void onLoad() {
+        EarlyLoadingManager.scan(this.getClass(), this);
     }
 
     @Override